Datamodel update test code
See https://github.com/PALEOtoolkit/PALEOmodel.jl/issues/48
https://github.com/PALEOtoolkit/PALEOboxes.jl/pull/85

- Generalize FieldArray
- move get_region and coordinate manipulation code from PALEOboxes.jl
- rename FixedCoord -> RecordCoord as now only used by
 FieldRecord (and should be used by OutputWriter)

TODO this currently allows a model (only the black sea config tested)
to run and store output, but the FieldRecord etc (and surely FieldArray
and plotting) does not work.
Stop at this point as this illustrates a review of the data model
for internal data storage (at least) is needed.
Which requires implementing a standardised stable output format on disk
first.